Add V2 of the panel
This codes almost all logic using a lambda.
Specifically the lambda is a lambda effect for the LEDs, this allows
it easy access to the LED components.

By using a lambda, it's much more intuitive, and allows for a new feature
which causes bright white flashing of the relevent LED when it changes state
to clearly indicate the change.